Token en /seguimiento_usuarios y algunos endpoints de /productos

+++ project/routes/productos.py
@@ -1,6 +1,7 @@
 from typing import List
 
 from fastapi import APIRouter
+from fastapi import Depends
 from fastapi import HTTPException
 
 from ..database import Producto
@@ -10,26 +11,32 @@ from ..schemas import ProductoRequestModel
 from ..schemas import ProductoResponseModel
 from ..schemas import ProductoRequestPutModel
 
+from ..middleware import get_current_user
+
 router = APIRouter(prefix='/productos')
 
 @router.post('', response_model=ProductoResponseModel)
-async def crear_producto(producto: ProductoRequestModel):
+async def crear_producto(producto: ProductoRequestModel, token: str = Depends(get_current_user)):
+
+    if token.get("nivel") == 1:
     
-    if Producto.select().where(Producto.sku == producto.sku).exists():
-        raise HTTPException(status_code=409, detail='Este SKU ya existe en el catálogo.')
+        if Producto.select().where(Producto.sku == producto.sku).exists():
+            raise HTTPException(status_code=409, detail='Este SKU ya existe en el catálogo.')
 
-    if Marca.select().where(Marca.id == producto.marca_id).first() is None:
-        raise HTTPException(status_code=404, detail='Marca no encontrada.')
+        if Marca.select().where(Marca.id == producto.marca_id).first() is None:
+            raise HTTPException(status_code=404, detail='Marca no encontrada.')
 
-    producto = Producto.create(
-        marca_id = producto.marca_id,
-        sku = producto.sku,
-        nombre = producto.nombre,
-        precio = producto.precio,
-        estatus = producto.estatus
-    )
+        producto = Producto.create(
+            marca_id = producto.marca_id,
+            sku = producto.sku,
+            nombre = producto.nombre,
+            precio = producto.precio,
+            estatus = producto.estatus
+        )
 
-    return producto
+        return producto
+
+    raise HTTPException(status_code=404, detail="Este usuario no tiene permisos para la petición.")
 
 @router.get('', response_model=List[ProductoResponseModel])
 async def obtener_productos(page: int = 1, limit : int = 10):
@@ -49,31 +56,39 @@ async def obtener_producto_id(producto_id: int):
     return producto_id
 
 @router.put('/{producto_id}', response_model=ProductoResponseModel)
-async def actualizar_producto(producto_id: int, review_request: ProductoRequestPutModel):
+async def actualizar_producto(producto_id: int, review_request: ProductoRequestPutModel, token: str = Depends(get_current_user)):
 
-    producto_id = Producto.select().where(Producto.id == producto_id).first()
+    if token.get("nivel") == 1:
 
-    if producto_id is None:
-        raise HTTPException(status_code=404, detail='Producto no encontrado.')
+        producto_id = Producto.select().where(Producto.id == producto_id).first()
 
-    producto_id.marca_id = review_request.marca_id
-    producto_id.sku = review_request.sku
-    producto_id.nombre = review_request.nombre
-    producto_id.precio = review_request.precio
-    producto_id.estatus = review_request.estatus
+        if producto_id is None:
+            raise HTTPException(status_code=404, detail='Producto no encontrado.')
 
-    producto_id.save()
+        producto_id.marca_id = review_request.marca_id
+        producto_id.sku = review_request.sku
+        producto_id.nombre = review_request.nombre
+        producto_id.precio = review_request.precio
+        producto_id.estatus = review_request.estatus
 
-    return producto_id
+        producto_id.save()
+
+        return producto_id
+
+    raise HTTPException(status_code=404, detail="Este usuario no tiene permisos para la petición.")
 
 @router.delete('/{producto_id}', response_model=ProductoResponseModel)
-async def eliminar_un_producto(producto_id: int):
+async def eliminar_un_producto(producto_id: int, token: str = Depends(get_current_user)):
 
-    producto_id = Producto.select().where(Producto.id == producto_id).first()
+    if token.get("nivel") == 1:
 
-    if producto_id is None:
-        raise HTTPException(status_code=404, detail='Producto no encontrado.')
-    
-    producto_id.delete_instance()
+        producto_id = Producto.select().where(Producto.id == producto_id).first()
 
-    return producto_id
+        if producto_id is None:
+            raise HTTPException(status_code=404, detail='Producto no encontrado.')
+        
+        producto_id.delete_instance()
+
+        return producto_id
+
+    raise HTTPException(status_code=404, detail="Este usuario no tiene permisos para la petición.")
